Clean up some lines in Intro flow
